requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.7
celery==5.3.4
redis==4.5.5
selenium==4.9.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
